python_graph/requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
scipy==1.16.3
six==1.17.0
